Refactor ColumnFilter into separate file and add tests
